differentiate "full" and "blocked" machines
Note: This is my first time poking at Lua, and at Factorio modding, so
other than intent, anything here can be wrong.

Previously all machines that were stopped and had *some* output waiting
were counted as "full".

This meant machines that were lacking inputs but had some material still in
the output tray, either sitting or being transferred out were given a
yellow light instead of a red light.

This wasn't optimal, as "red" was easy to read as "materials missing", but
when reading "yellow" as "the output is bottlenecking the machine" was
sometimes correct and other times not.

After this change the "blocked" status is applied to machines that have a
completed crafting cycle yet aren't working. This is marked with "yellow"
and can be read with high confidence as "the output is bottlenecking the
machine".

Stopped machines without a full crafting cycle are now marked as stopped
or full separately, so one can still search for machines with stuff in them via the
icon, but both are "red" now so it's easy as "bottlenecked due to missing
inputs".