Redraw and simplify SVGs
The SVGs are now 16px x 16px with no padding to readily expand up to a 64px
square. The gradient has been removed to make the distinction between
different-colored dots more clear as well as to make the icons more readable.
The black stroke has been replaced with a slightly opaque one.

In addition, the SVG code has been simplified to better accomodate simple
changes to the graphics.